refactor(Core UI): :recycle: Move shared pages to core UI
