+++ __tests__/MetricLatency.spec.tsx
@@ -10,8 +10,8 @@ import * as PrometheusAPIModule from '../src/API/Prometheus.api';
 import { waitForElementToBeRemovedTimeout } from '../src/config/app';
 import { Labels } from '../src/config/labels';
 import { getTestsIds } from '../src/config/testIds';
+import Latency from '../src/core/components/Metrics/components/Latency';
 import LoadingPage from '../src/core/components/SkLoading';
-import Latency from '../src/pages/shared/Metrics/components/Latency';
 import { Providers } from '../src/providers';
 
 const processResult = processesData.results[0];
