+++ __tests__/MetricRequest.spec.tsx
@@ -10,8 +10,8 @@ import * as PrometheusAPIModule from '../src/API/Prometheus.api';
 import { waitForElementToBeRemovedTimeout } from '../src/config/app';
 import { Labels } from '../src/config/labels';
 import { getTestsIds } from '../src/config/testIds';
+import Request from '../src/core/components/Metrics/components/Request';
 import LoadingPage from '../src/core/components/SkLoading';
-import Request from '../src/pages/shared/Metrics/components/Request';
 import { Providers } from '../src/providers';
 
 const processResult = processesData.results[0];
