+++ __tests__/SkNavigationViewLink.spec.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { describe, expect, it } from 'vitest';
 
-import SkNavigationViewLink from '../src/core/components/SkNavigationViewLink';
+import SkNavigationViewLink from '../src/core/components/SkTable/SkNavigationViewLink';
 
 describe('NavigationViewLink', () => {
   it('renders the link with the correct label and default icon', () => {
