+++ src/App.css
@@ -1,3 +1,5 @@
-html, body, #app {
+html,
+body,
+#app {
   height: 100%;
 }
