+++ src/core/components/Metrics/components/ResponseCharts.tsx
