+++ src/core/components/Metrics/hooks/useMetricsSessionHandler.ts
