+++ src/pages/Processes/components/Overview.tsx
@@ -1,10 +1,10 @@
 import { FC } from 'react';
 
 import { Protocols } from '../../../API/REST.enum';
+import Metrics from '../../../core/components/Metrics';
 import { extractUniqueValues } from '../../../core/utils/extractUniqueValues';
 import { mapDataToMetricFilterOptions } from '../../../core/utils/getResourcesFromPairs';
 import { ProcessResponse } from '../../../types/REST.interfaces';
-import Metrics from '../../shared/Metrics';
 import { useProcessOverviewData } from '../hooks/useOverviewData';
 
 interface OverviewProps {
