+++ src/pages/Services/components/Overview.tsx
@@ -1,9 +1,9 @@
 import { FC } from 'react';
 
 import { Protocols } from '../../../API/REST.enum';
+import Metrics from '../../../core/components/Metrics';
 import { extractUniqueValues } from '../../../core/utils/extractUniqueValues';
 import { mapDataToMetricFilterOptions } from '../../../core/utils/getResourcesFromPairs';
-import Metrics from '../../shared/Metrics';
 import { useServiceOverviewData } from '../hooks/useOverviewData';
 
 interface OverviewProps {
