+++ src/pages/Services/components/TcpConnections.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 
+import BiFlowLogs from '../../../core/components/BiFlowLogs';
 import { tcpBiFlowColumns, tcpSelectOptions } from '../../../core/components/SkBiFlowList/BiFlowList.constants';
 import { setColumnVisibility } from '../../../core/components/SkTable/SkTable.utils';
-import BiFlowLogs from '../../shared/BiFlowLogs';
 import { initActiveConnectionsQueryParams } from '../Services.constants';
 
 interface TcpConnectionsProps {
