+++ src/pages/Sites/components/Overview..tsx
@@ -1,11 +1,11 @@
 import { FC } from 'react';
 
 import { Protocols } from '../../../API/REST.enum';
+import Metrics from '../../../core/components/Metrics';
 import { extractUniqueValues } from '../../../core/utils/extractUniqueValues';
 import { mapDataToMetricFilterOptions } from '../../../core/utils/getResourcesFromPairs';
 import { removeDuplicatesFromArrayOfObjects } from '../../../core/utils/removeDuplicatesFromArrayOfObjects';
 import { SiteResponse } from '../../../types/REST.interfaces';
-import Metrics from '../../shared/Metrics';
 import { useSiteOverviewData } from '../hooks/useOverviewData';
 
 interface OverviewProps {
