+++ src/routes/errorRoutes.tsx
@@ -1,8 +1,8 @@
 import { lazy } from 'react';
 
-import { NotFoundRoutesPaths } from '../pages/shared/Errors/NotFound/NotFound.enum';
+import { NotFoundRoutesPaths } from '../core/components/Errors/NotFound/NotFound.enum';
 
-const NotFound = lazy(() => import(/* webpackChunkName: "not-found-error" */ '../pages/shared/Errors/NotFound'));
+const NotFound = lazy(() => import(/* webpackChunkName: "not-found-error" */ '../core/components/Errors/NotFound'));
 
 export const errorsRoutes = [
   {
